refactor: 필요없는 설정파일 삭제
