--- src/main/java/com/kustacks/kuring/common/dto/ResponseDto.java
@@ -1,19 +1,0 @@
-package com.kustacks.kuring.common.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class ResponseDto {
-
-    @JsonProperty("isSuccess")
-    private boolean success;
-
-    private String resultMsg;
-
-    private int resultCode;
-}
