--- src/main/java/com/kustacks/kuring/config/HttpSessionConfig.java
@@ -1,27 +1,0 @@
-package com.kustacks.kuring.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.session.MapSessionRepository;
-import org.springframework.session.config.annotation.web.http.EnableSpringHttpSession;
-import org.springframework.session.web.http.CookieSerializer;
-import org.springframework.session.web.http.DefaultCookieSerializer;
-
-import java.util.concurrent.ConcurrentHashMap;
-
-@EnableSpringHttpSession
-public class HttpSessionConfig {
-
-    @Bean
-    public MapSessionRepository sessionRepository() {
-        return new MapSessionRepository(new ConcurrentHashMap<>());
-    }
-
-    @Bean
-    public CookieSerializer cookieSerializer() {
-        DefaultCookieSerializer serializer = new DefaultCookieSerializer();
-        serializer.setCookieName("KURING");
-        serializer.setCookiePath("/admin");
-        serializer.setCookieMaxAge(60 * 120);
-        return serializer;
-    }
-}
