+++ src/main/java/com/kustacks/kuring/worker/scrap/StaffScraper.java
@@ -1,6 +1,6 @@
 package com.kustacks.kuring.worker.scrap;
 
-import com.kustacks.kuring.common.dto.StaffDto;
+import com.kustacks.kuring.staff.common.dto.StaffDto;
 import com.kustacks.kuring.common.exception.code.ErrorCode;
 import com.kustacks.kuring.common.exception.InternalLogicException;
 import com.kustacks.kuring.worker.scrap.client.staff.StaffApiClient;
